requirements.txt
opencv-python
PyQt5
matplotlib
pygame
ultralytics
pillow
numpy
pyserial
